Handle Sqlite operations on derived covariant types
If you are running a BulkInsert from a method which accepts AbstractBaseClass, and pass in a set of entities which are DerivedClass, SQL server operations correctly use the derived class entity configuration (ToTable, etc.). However when using Sqlite, it attempts to use the properties of the base class in the type accessor, which will not work if you have extra properties on your derived class.